+++ src/Map_JS.js
@@ -1,13 +1,15 @@
 import Fuse from "fuse.js";
 import initModal from "./modal";
 import "./map";
+import * as cart from "./cart";
+import makeDataTable from "./dataTable";
 
 ("use strict");
 
 /* READ IN LOCAL JSON AS ARRAY */
 fetch("DocsJSON.json")
-  .then(res => res.json())
-  .then(data => {
+  .then((res) => res.json())
+  .then((data) => {
     writeData(data); // writes JSON as JS object
     makeDataTable(data); // presents data in table
     fuseListener(); // listener for changes to search field
@@ -15,153 +17,51 @@ fetch("DocsJSON.json")
     additionListener(); // listener for doc adds
     removalListener(); // listener for doc removes
     initModal(); // initiates modal functions
-    makeCartTable(); // generates shopping cart table in modal
+    cart.makeCartTable(); // generates shopping cart table in modal
   });
 
 let data;
 
-const writeData = source => {
+const writeData = (source) => {
   data = source;
-};
-
-/* LOCAL JSON TABLE */
-const makeDataTable = data => {
-  let html = '<table class="table is-striped" id="addition">';
-  html += "<tr>";
-  const headers = ["Year", "Author", "Title", ""];
-  headers.forEach(header => {
-    return (html += `<th>${header}</th>`);
-  });
-  html += "</tr>";
-  data.forEach(doc => {
-    const tableRow = `<tr class="addCart">
-      <td>${doc.Year}</td>
-      <td>${doc.Author}</td>
-      <td>${doc.Title}</td>
-      <td><button class="button is-link is-outlined is-small addCart" id="${doc.FID}">Add to cart</button></td>
-      </tr>`;
-    return (html += tableRow);
-  });
-  html += "</table>";
-  document.querySelector("div#clicky").innerHTML = html;
 };
 
 /* SHOPPING CART */
 // If user has cart data in their localStorage, this becomes `docsInCart`
-let cartStore;
-let docsInCart;
 
-const initCartStore = () => {
-  if (JSON.parse(window.localStorage.getItem("cart")) === null) {
-    window.localStorage.setItem("cart", "[]");
-  }
-  cartStore = JSON.parse(window.localStorage.getItem("cart"));
-  docsInCart = cartStore;
-};
-
-const modifyCartStore = cart => {
-  window.localStorage.setItem("cart", JSON.stringify(cartStore));
-  cartStore = JSON.parse(window.localStorage.getItem("cart"));
-};
-
-function addDoc(fid, source, cart) {
-  source.forEach(doc => {
-    if (fid === doc.FID) {
-      if (cart && cart.length) {
-        let found = false;
-
-        for (let i = 0; i < cart.length; i++) {
-          if (cart[i].FID === doc.FID) {
-            found = true;
-            break;
-          }
-        }
-
-        if (found === false) {
-          cart.push(doc);
-        }
-      } else {
-        cart.push(doc);
-      }
-    }
-  });
-  modifyCartStore(cart);
-  makeCartTable();
-}
-
-function removeDoc(fid, cart) {
-  let docIndex;
-
-  cart.forEach(doc => {
-    if (fid === doc.FID) {
-      docIndex = cart.indexOf(doc);
-    }
-  });
-  cart.splice(docIndex, 1);
-  modifyCartStore(cart);
-  makeCartTable();
-}
-
-const additionListener = () => {
+export const additionListener = () => {
   const addition = document.getElementById("clicky");
   addition.addEventListener(
     "click",
-    event => {
+    (event) => {
       event.preventDefault();
       const isButton = event.target.nodeName === "BUTTON";
       if (!isButton) {
         return;
       }
       let fid = event.target.id;
-      addDoc(fid, data, docsInCart);
+      cart.addDoc(fid, data, cart.docsInCart);
     },
     false
   );
 };
 
-const removalListener = () => {
+export const removalListener = () => {
   const removal = document.getElementById("cart");
   removal.addEventListener(
     "click",
-    event => {
+    (event) => {
       event.stopPropagation();
       const isButton = event.target.nodeName === "BUTTON";
       if (!isButton) {
         return;
       }
       let fid = event.target.id;
-      removeDoc(fid, docsInCart);
+      cart.removeDoc(fid, cart.docsInCart);
     },
     false
   );
 };
-
-function makeCartTable() {
-  if (!cartStore) {
-    docsInCart = [];
-  } else {
-    docsInCart = cartStore;
-  }
-
-  let html = '<table class="table is-striped">';
-  html += "<tr>";
-  const headers = ["Year", "Author", "Title", ""];
-  headers.forEach(header => {
-    return (html += `<th>${header}</th>`);
-  });
-  html += "</tr>";
-  cartStore.forEach(doc => {
-    const tableRow = `<tr>
-      <td>${doc.Year}</td>
-      <td>${doc.Author}</td>
-      <td>${doc.Title}</td>
-      <td><button class="button is-danger is-outlined is-small" id="${doc.FID}">Remove from cart</button></td>
-      </tr>`;
-    return (html += tableRow);
-  });
-  html += "</table>";
-  document.querySelector("div#cart").innerHTML = html;
-}
 
 /* FUSE SEARCH */
 let fuse;
@@ -177,13 +77,13 @@ let options = {
   distance: 100,
   maxPatternLength: 32,
   minMatchCharLength: 1,
-  keys: ["Title", "Author", "contentTags"]
+  keys: ["Title", "Author", "contentTags"],
 };
 
 /* FUSE SEARCH RESULTS */
 const fuseListener = () => {
   const fusefield = document.getElementById("fusefield");
-  fusefield.addEventListener("keyup", event => {
+  fusefield.addEventListener("keyup", (event) => {
     event.stopPropagation();
     fusesearch(fusefield);
   });
@@ -204,11 +104,11 @@ function fusesearch(fusefield) {
     html += "<tr>";
     var flag = 0;
     var headers = ["Year", "Author", "Title", ""];
-    headers.forEach(header => {
+    headers.forEach((header) => {
       return (html += `<th>${header}</th>`);
     });
     html += "</tr>";
-    result.forEach(doc => {
+    result.forEach((doc) => {
       const tableRow = `<tr>
       <td>${doc.item.Year}</td>
       <td>${doc.item.Author}</td>
@@ -252,15 +152,15 @@ function getInputs(form) {
 
 /* LISTENER FOR CLICKING SUBMIT BUTTON */
 
-document.addEventListener("DOMContentLoaded", function() {
+document.addEventListener("DOMContentLoaded", function () {
   const form = document.getElementById("requestForm");
   const output = document.getElementById("output");
 
-  initCartStore();
+  cart.initCartStore();
 
   form.addEventListener(
     "submit",
-    function(e) {
+    function (e) {
       e.preventDefault();
       let user = getInputs(this);
       let docs = wrapDocs();
@@ -273,11 +173,11 @@ document.addEventListener("DOMContentLoaded", function() {
       formData.append("request", tmp);
 
       fetch("./api/sg-trans.php", { method: "POST", body: formData })
-        .then(function(response) {
+        .then(function (response) {
           window.localStorage.removeItem("cart");
           return response.text();
         })
-        .then(function(body) {
+        .then(function (body) {
           console.log(body);
         });
     },
@@ -287,7 +187,7 @@ document.addEventListener("DOMContentLoaded", function() {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-document.onreadystatechange = function() {
+document.onreadystatechange = function () {
   if (document.readyState === "complete") {
     function setCoordinates() {
       var Long = document.getElementById("Long").value;
